add support for key denylist
This is to filter out cluster rows where the resulting key is in a given
text file (one key per line).

The intent is to filter out records with either poor metadata, or very
generic metadata, for fuzzy matching. Eg, in many cases it is better to
just not try matching "Letter to the Editor" to any record. This won't
always be the case; we might have journal, volume, issue, and page,
which would allow a match. So this can be specified on the command line.